% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Descriptive statistics}
\usage{
describe(data, by, detailed = FALSE, ...)
}
\arguments{
\item{data}{= a matrix or a data.frame}

\item{by}{= the name of a variable to condition on}

\item{detailed}{= logical, if TRUE a detailed summary is returned}
}
\description{
Descriptive statistics for a matrix or data frame.
}
\examples{

describe(iris, detailed = TRUE)
describe(iris, by = "Species")

}
