% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisherz.R
\name{FisherZ}
\alias{FisherZ}
\alias{FisherZinv}
\title{Fisher Z transformation}
\usage{
FisherZ(rho)

FisherZinv(z)
}
\arguments{
\item{rho}{= the value of the correlation in \eqn{[0,1]}}

\item{z}{= the value of the z-transform in \eqn{(-\infty, +\infty)}}
}
\value{
The transformed or back-transformed value.
}
\description{
Computes the Fisher Z-transform and its inverse.
}
\examples{

FisherZ(0.8)
FisherZinv(FisherZ(-0.5))

}
