% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fscoreSummary.R
\name{fscoreSummary}
\alias{fscoreSummary}
\title{Precision-Recall-Fscore performance measures across resamples}
\usage{
fscoreSummary(data, lev = NULL, model = NULL, ...)
}
\arguments{
\item{data}{a data frame with columns \code{obs} and \code{pred} for the 
observed and predicted outcomes, and \code{prob} for predicted probabilities 
for each class. 
See the \code{classProbs} argument to \code{\link[caret]{trainControl}}.}

\item{lev}{a character vector of factors levels for the response.}

\item{model}{a character string for the model name (as taken from the 
\code{method} argument of \code{\link[caret]{train}}.}
}
\value{
A vector containing the metrics.
}
\description{

}
\details{
\code{\link{https://en.wikipedia.org/wiki/Precision_and_recall}}
}
\examples{
data = caret::twoClassSim(200)

mod1 = train(Class ~ . , data = data,
             method = "rpart2",
             tuneGrid = expand.grid(maxdepth = 1:10),
             metric = "F1",
             trControl = trainControl(method = "cv", number = 10,
                                      classProbs = TRUE,
                                      savePredictions = TRUE,
                                      summaryFunction = fscoreSummary,
                                      selectionFunction = "best") )
mod1
plot(mod1)

mod2 = train(Class ~ . , data = data,
             method = "rpart2",
             tuneGrid = expand.grid(maxdepth = 1:10),
             metric = "F2",
             trControl = trainControl(method = "cv", number = 10,
                                      classProbs = TRUE,
                                      summaryFunction = fscoreSummary,
                                      selectionFunction = "best") )
mod2
plot(mod2)

mod3 = train(Class ~ . , data = data,
             method = "rf",
             tuneGrid = expand.grid(mtry = 1:5),
             metric = "F1",
             trControl = trainControl(method = "cv", number = 10,
                                      classProbs = TRUE,
                                      summaryFunction = fscoreSummary,
                                      selectionFunction = "best") )
mod3
plot(mod3)

mod4 = train(Class ~ . , data = data,
             method = "rf",
             tuneGrid = expand.grid(mtry = 1:5),
             metric = "F2",
             trControl = trainControl(method = "cv", number = 10,
                                      classProbs = TRUE,
                                      summaryFunction = fscoreSummary,
                                      selectionFunction = "best") )
mod4
plot(mod4)


}
\seealso{
\code{\link[caret]{twoClassSummary}}
}
