% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Wald confidence intervals for Maximum Likelihood Estimates}
\usage{
plotImage(x, max = 255, mar = c(0, 0, 0, 0), ...)
}
\arguments{
\item{x}{a matrix of data values for grey-scale images and a 3-dimensional array for RGB images.}

\item{max}{the maximum value of RGB color scale.}

\item{mar}{the margin used in the base R graphic.}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\description{
Wald confidence intervals based on asymptotic Gaussian
distribution of MLE.
}
