% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_addson.R
\name{prcompRecon}
\alias{prcompRecon}
\title{Principal components data reconstruction}
\usage{
prcompRecon(object, npcs = 2, ...)
}
\arguments{
\item{object}{= an object returned by 'prcomp' function.}

\item{npcs}{= the number of principal componets to use in the reconstruction}
}
\value{
The reconstructed data matrix.
}
\description{
Return the data matrix reconstructed using a subset of 
principal components.
}
\examples{

pairs(iris[,1:4], gap = 0)
PCA = prcomp(iris[,1:4])
X = prcompRecon(PCA, 2)
pairs(X, gap = 0)

}
