% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_addson.R
\name{prcompReconError}
\alias{prcompReconError}
\title{Principal components reconstruction error}
\usage{
prcompReconError(object, data, ...)
}
\arguments{
\item{object}{an object returned by 'prcomp' function.}

\item{data}{the original data matrix}

\item{\dots}{additional arguments to be passed to the low level functions.}
}
\value{
A data frame containing the reconstruction error (RE), the mean
square error (MSE), and the R^2 for increasing value of principal 
components.
}
\description{
Return the squared reconstruction error from using a subset of 
principal components.
}
\examples{

PCA = prcomp(iris[,1:4])
df = prcompReconError(PCA, iris[,1:4])
zapsmall(df,5)

}
