% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatterplot with marginal histograms}
\usage{
scatterhist(x, y = NULL, ...)
}
\arguments{
\item{x}{a vector, matrix or data frame}

\item{y}{if provided, a vector}
}
\value{
None, just produces a plot.
}
\description{
Plot a scatterplot of bivariate data with histograms for the marginal distributions.
}
\examples{

x <- rnorm(200)
y <- 0.5*x + rnorm(200, 0, 0.1)
scatterhist(x, y)
}
