% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Plot image data}
\usage{
plotImage(x, max = 255, mar = c(0, 0, 0, 0), restore.par = TRUE, ...)
}
\arguments{
\item{x}{a matrix of data values for Grey-scale images and a 
3-dimensional array for RGB images.}

\item{max}{the maximum value of Grey/RGB colour scale (default 255 
for integer values in the range [0, 255]; another common value is 1 
for values in the range [0, 1].}

\item{mar}{the margin used in the base R graphic.}

\item{restore.par}{logical, if TRUE the previous par() of graphical 
window are restored.}

\item{\dots}{additional arguments to be passed to the low level 
functions.}
}
\description{
Plot a matrix representing a Grey-scale image or a 
3-dimensional array representing a RGB image.
}
